+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.0.0-beta.90
+
+### instant-apps-interactive-legend
+
+- Add support for up to three nested groups.
+- Re-adjust line clamp back to 2.
+
 ## v1.0.0-beta.89
 
 ### instant-apps-interactive-legend